moved array stuff to helper - tested
